+++ src/Configuration/Cache/Memcached.php
@@ -25,6 +25,7 @@ class Memcached extends AbstractCache
      */
     protected $keys = [
         'memcached_dsn',
+        'use_cache_tags'
     ];
 
     public function getConfigTreeBuilder(): TreeBuilder
@@ -36,7 +37,24 @@ class Memcached extends AbstractCache
             ->scalarNode('memcached_dsn')->isRequired()->cannotBeEmpty()->end()
         ->end()
         ;
+        $this->addCommonNodes($rootNode);
+        $this->validate($rootNode);
 
         return $treeBuilder;
+    }
+
+    /**
+     * Conditional validation.
+     *
+     * @return void
+     */
+    protected function validate($rootNode)
+    {
+        $rootNode->validate()
+            ->ifTrue(function (array $v) {
+                return true === $v['use_cache_tags'];
+            })
+            ->thenInvalid('Cache tags is not supported by Memcached cache.')
+        ->end();
     }
 }